Re-locate MyFriend application button
Make MyFriend application button to be fixed on the bottom.
