+++ src/locales.json
@@ -190,14 +190,15 @@
     "microcare_service2_title": "병원 통역 서비스",
     "microcare_title": "간편하게 카톡하세요",
     "myfriend_caution": "### 유의사항\n\n1. 병원에서 통역을 허가하지 않으면 서비스가 제공되지 않습니다.\n1. 휴대폰의 스피커폰 기능을 통해 통역이 이루어지기 때문에 개인 휴대폰의 성능에 대해 사전 확인이 필요합니다.\n1. 통역은 대화 내용만을 전달하는 것으로 통역범위를 넘어서는 요구는 제한 될 수 있습니다.\n1. 불법과 관련된 통역은 서비스에 포함되지 않습니다.\n1. 고객 민감정보는(증상, 병명등) 어떠한 형태로도 보관되지 않고 오로지 원활한 통역을 위해서만 사용됩니다.",
-    "myfriend_description": "## 독일에서 한국어로 진료 받으세요\n\n4만 명이 넘는 한국인이 독일에 유학, 이민, 근로 등의 목적으로 거주하고 있지만,  \n생각보다 독일어는 어렵습니다.\n\n특히, 병원에서 의사소통 때문에 답답한 상황에 처하기도하고  \n심지어 아파도 병원에 가지 않는 경우가 있다는 것을 알았습니다.\n\n마이크로프로텍트가 무료통역으로 도와드리겠습니다.\n\n저희는 유럽에 거주하는 한국인을 위한 보험 서비스를 제공하는 회사로서,  \n교민들의 불편함을 해결하려는 사명을 가지고 있습니다.\n\n마치 가까운 친구에게 부탁하듯, 마이프렌드 무료 통역을 이용하세요.\n\n> ### 마이프렌드 무료통역 서비스 운영 시간\n>\n> 평일 09:00~17:00 (토요일, 공휴일 제외)",
+    "myfriend_description": "4만 명이 넘는 한국인이 독일에 유학, 이민, 근로 등의 목적으로 거주하고 있지만,  \n생각보다 독일어는 어렵습니다.\n\n특히, 병원에서 의사소통 때문에 답답한 상황에 처하기도하고  \n심지어 아파도 병원에 가지 않는 경우가 있다는 것을 알았습니다.\n\n마이크로프로텍트가 무료통역으로 도와드리겠습니다.\n\n저희는 유럽에 거주하는 한국인을 위한 보험 서비스를 제공하는 회사로서,  \n교민들의 불편함을 해결하려는 사명을 가지고 있습니다.\n\n마치 가까운 친구에게 부탁하듯, 마이프렌드 무료 통역을 이용하세요.\n\n> ### 마이프렌드 무료통역 서비스 운영 시간\n>\n> 평일 09:00~17:00 (토요일, 공휴일 제외)",
     "myfriend_guide_step1": "카카오톡에서\n마이프렌드 채널을 추가하세요",
     "myfriend_guide_step2": "마이프렌드 채널에서\n온라인 신청서를 작성하세요",
     "myfriend_guide_step3": "신청 후 상담사와 세부 사항을 논의합니다",
     "myfriend_guide_step4": "담당 통역사가 스피커폰을 통해\n병원에서 소통을 도와드립니다",
     "myfriend_guide_title": "이용 가이드",
     "myfriend_mobile_only": "예약은 모바일로만 가능합니다",
-    "myfriend_request_button": "병원 통역 신청하기",
+    "myfriend_request_button": "무료통역 신청하기",
+    "myfriend_slogan": "독일에서 한국어로 진료 받으세요",
     "myfriend_title": "My Friend",
     "partner_description_busan": "We are part of a team that fosters and supports the Busan Creative Economy Innovation Center.",
     "partner_description_d3": "D3 Jubilee Partners is a pioneer in social impact investing in Korea.\r\nD3 has played a big role in enabling Microprotect to do insurance business in Germany and to collaborate with Balance Hero in India.",
@@ -441,14 +442,15 @@
     "microcare_service2_title": "병원 통역 서비스",
     "microcare_title": "간편하게 카톡하세요",
     "myfriend_caution": "### 유의사항\n\n1. 병원에서 통역을 허가하지 않으면 서비스가 제공되지 않습니다.\n1. 휴대폰의 스피커폰 기능을 통해 통역이 이루어지기 때문에 개인 휴대폰의 성능에 대해 사전 확인이 필요합니다.\n1. 통역은 대화 내용만을 전달하는 것으로 통역범위를 넘어서는 요구는 제한 될 수 있습니다.\n1. 불법과 관련된 통역은 서비스에 포함되지 않습니다.\n1. 고객 민감정보는(증상, 병명등) 어떠한 형태로도 보관되지 않고 오로지 원활한 통역을 위해서만 사용됩니다.",
-    "myfriend_description": "## 독일에서 한국어로 진료 받으세요\n\n4만 명이 넘는 한국인이 독일에 유학, 이민, 근로 등의 목적으로 거주하고 있지만,  \n생각보다 독일어는 어렵습니다.\n\n특히, 병원에서 의사소통 때문에 답답한 상황에 처하기도하고  \n심지어 아파도 병원에 가지 않는 경우가 있다는 것을 알았습니다.\n\n마이크로프로텍트가 무료통역으로 도와드리겠습니다.\n\n저희는 유럽에 거주하는 한국인을 위한 보험 서비스를 제공하는 회사로서,  \n교민들의 불편함을 해결하려는 사명을 가지고 있습니다.\n\n마치 가까운 친구에게 부탁하듯, 마이프렌드 무료 통역을 이용하세요.\n\n> ### 마이프렌드 무료통역 서비스 운영 시간\n>\n> 평일 09:00~17:00 (토요일, 공휴일 제외)",
+    "myfriend_description": "4만 명이 넘는 한국인이 독일에 유학, 이민, 근로 등의 목적으로 거주하고 있지만,  \n생각보다 독일어는 어렵습니다.\n\n특히, 병원에서 의사소통 때문에 답답한 상황에 처하기도하고  \n심지어 아파도 병원에 가지 않는 경우가 있다는 것을 알았습니다.\n\n마이크로프로텍트가 무료통역으로 도와드리겠습니다.\n\n저희는 유럽에 거주하는 한국인을 위한 보험 서비스를 제공하는 회사로서,  \n교민들의 불편함을 해결하려는 사명을 가지고 있습니다.\n\n마치 가까운 친구에게 부탁하듯, 마이프렌드 무료 통역을 이용하세요.\n\n> ### 마이프렌드 무료통역 서비스 운영 시간\n>\n> 평일 09:00~17:00 (토요일, 공휴일 제외)",
     "myfriend_guide_step1": "카카오톡에서\n마이프렌드 채널을 추가하세요",
     "myfriend_guide_step2": "마이프렌드 채널에서\n온라인 신청서를 작성하세요",
     "myfriend_guide_step3": "신청 후 상담사와 세부 사항을 논의합니다",
     "myfriend_guide_step4": "담당 통역사가 스피커폰을 통해\n병원에서 소통을 도와드립니다",
     "myfriend_guide_title": "이용 가이드",
     "myfriend_mobile_only": "예약은 모바일로만 가능합니다",
-    "myfriend_request_button": "병원 통역 신청하기",
+    "myfriend_request_button": "무료통역 신청하기",
+    "myfriend_slogan": "독일에서 한국어로 진료 받으세요",
     "myfriend_title": "마이프렌드",
     "partner_description_busan": "정부에서 진행하는 사업인 부산창조경제혁신센터에서 육성하고 지원하는 팀에 소속되어 있습니다. ",
     "partner_description_d3": "D3 쥬빌리 파트너스는 한국의 소셜 임팩트 투자에서 선구자의 역할을 하고 있습니다.\r\nD3는 마이크로프로텍트가 독일에서 보험사업을 할 수 있도록, 그리고 인도에서 밸런스 히어로와 협업을 할 수 있도록 큰 역할을 해주셨습니다. ",
